+++ src/Serilog.Sinks.AzureTableStorage/LoggerConfigurationAzureTableStorageExtensions.cs
@@ -13,8 +13,7 @@
 // limitations under the License.
 
 using System;
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Auth;
+using Microsoft.Azure.Cosmos.Table;
 using Serilog.Configuration;
 using Serilog.Core;
 using Serilog.Events;
@@ -334,7 +333,7 @@ namespace Serilog
                 }
                 else
                 {
-                    storageAccount = new CloudStorageAccount(credentials, null, null, tableEndpoint, null);
+                    storageAccount = new CloudStorageAccount(credentials, tableEndpoint);
                 }
 
                 // We set bypassTableCreationValidation to true explicitly here as the the SAS URL might not have enough permissions to query if the table exists.