+++ src/Serilog.Sinks.AzureTableStorage/Sinks/AzureTableProvider/ICloudTableProvider.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Table;
+using Microsoft.Azure.Cosmos.Table;
 
 namespace Serilog.Sinks.AzureTableStorage.AzureTableProvider
 {