+++ src/Serilog.Sinks.AzureTableStorage/Sinks/AzureTableStorageWithProperties/AzurePropertyFormatter.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Microsoft.WindowsAzure.Storage.Table;
+using Microsoft.Azure.Cosmos.Table;
 using Serilog.Events;
 using System;