+++ CHANGES.md
@@ -1,6 +1,6 @@
 5.0.0
  * Migrated to new CSPROJ project system
- * Updated dependencies: WindowsAzure.Storage 8.6.0, Serilog 2.6.0, Serilog.Sinks.PeriodicBatching 2.1.1
+ * Updated dependencies: Microsoft.Azure.Cosmos.Table 1.0.6, Serilog 2.9.0, Serilog.Sinks.PeriodicBatching 2.3.0
  * Fix #36 - Allow using SAS URI for logging.
 
 1.5